Outsource Form to component
